second commit

+++ src/com/dsy/ActionBoneParse/util/Graphics.java
@@ -565,7 +565,7 @@ public final class Graphics implements UserData {
 	}
 
 	/* drawLine */
-	public void drawLine(int x1, int y1, int x2, int y2) {
+	public void drawLine(float x1, float y1, float x2, float y2) {
 		canvas.drawLine(x1, y1, x2, y2, paint);
 	}
